Add new shadow style TRANSPARENT_SHADOWED
It is the same as SHADOWED except when the object is completely
transparent, then the shadowed the part will override the transparency
to what's set in SoShadowTransparencyElement::transparency. Node
SoShadowTransparency can be used to customize the tranparency.

For the current impelementation, if you want to customize shadow
transparency, you must put the SoShadowTransparency node before the
SoShadowStyle node with the TRANSPARENT_SHADOWED style.