Refactor smoothBorder in SoShadowGroup
The original implementation uses gaussian blur for pre-filtering the
shadow map. The code has a few problems. The commit fixed it and
greatly improved performance by using separable filtering, i.e. a two
pass filtering in horizontal and vertical direction. The first pass
reads the shadow map texture, and outputs to gaussmap texture. The
second pass reads the gaussmap texture, and output back the shadow map.
(reference: http://rastergrid.com/blog/2010/09/efficient-gaussian-blur-with-linear-sampling/)

A future improvement may replace the gaussian blue with'Summed Area Table'.
http://www.shaderwrangler.com/publications/sat/SAT_EG2005.pdf

'smoothBorder' now uses the integer and the first fractional digits to
control the filter radius.

In addition to pre-filtering with gaussian blue, which is mostly only
for border smoothing, a post-filter is also implemented in this patch
that filters while doing shadow rendering. It is mostly used for soft
shadow rendering. The implementation is inspired by the book 'Computer
Graphics Programming in OpenGL with C++' by V. Scott Gordon, John L.
Clevenger.

https://books.google.com.sg/books?id=GomCDwAAQBAJ&lpg=PT290&vq=soft%20shadow&pg=PT298#v=onepage&q&f=false

The sampling is done in shadow rendering fragment shader. The second
fractional digit of smoothBorder feild is used to control the sampling
size. When the digit is 0, the filter uses 2x2 samples with dithering.
From there onwards, digit*2+1 specifies a box size for box filtering,
e.g. 1 means a cube of 3x3, and 2 for 5x5, etc, up to maximum 8x8.

The third+ digits of smoothBorder determine the shadow spread i.e.
penumbra size. Note that in case of spot light, the penumbra size will
be scaled by perspective projection, which will make it bigger the
further the sahdow is from the light source, which is what real shadows
do.

There is a shortcoming of the current implementation. The shadow spread
is dependent on the texture matrix. In other words, the higher the scale
implied by texture mapping, the smaller the spread shall be used. A
spread that is too big will produce incorrect shadows.