Fix potential crash on out- of-ordered path list traversal
SoGLRenderAction traverse path list has some assumption about the
traversing order. It assume the paths that have common acenstor
branches is added in the same order as the traversing order.

SoGLRenderAction::apply(SoPathList*, bool) has a second argument to
inform if this assumption holds true. If the assumption is made but the
input path list violates the assumption, it may cause crash. And it is
fixed by this commit.